Improvements to parse visualization
It now shows the different steps in the parser phase, though the first
one is wrong.

The issue now is that d3's animation doesn't work for inserting nodes
into the middle of the tree, only on the ends. So, once the transition
is somewhat complex, the whole thing is reshown, and you don't get as
good of a sense of what's changing as I'd like.

However, I've managed to capture the steps and the difference in the
tree nicely.